% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural-nets.R
\docType{package}
\name{deep}
\alias{deep}
\alias{deep-package}
\title{deep: A Neural Networks Framework}
\description{
The deep package provides classes for layers, types of neurons and the
neural network as a whole.
}
