% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCullochPitts.R
\docType{class}
\name{McCullochPitts-class}
\alias{McCullochPitts-class}
\alias{mcCullochPitts}
\title{The McCullochPitts neuron class, that implements the logic of the
McCullochPitts neuron model.}
\arguments{
\item{inputs}{The actual data to be fed to the nuron, this input's
dimentions vary with the chosen weights dimentions.}

\item{ins}{The list of vectors of inputs to the first layer in the network}

\item{outs}{The list of vectors of outputs of the last layer in the network}

\item{epochs}{How many rounds of training to run}

\item{tax}{This is the learning rate, aka eta}

\item{maxErr}{A contition to early stop the training process}
}
\value{
The computed value using the McCullochPitts model.

Vector of computed values of the same size of the last layer
}
\description{
The McCullochPitts neuron class, that implements the logic of the
McCullochPitts neuron model.
}
\section{Fields}{

\describe{
\item{\code{ws}}{The matrix of weights that multiply the input vector, it can be a
vector, a matrix or an array.}

\item{\code{bias}}{The bias value.}
}}


