% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural-nets.R
\docType{class}
\name{NeuralNetwork-class}
\alias{NeuralNetwork-class}
\alias{neuralNet}
\title{The main NeuralNetwork class, that holds the layers.}
\description{
The main NeuralNetwork class, that holds the layers.
}
\section{Fields}{

\describe{
\item{\code{eta}}{The learning tax, representes the size of the weight adjustment
between each epoch of training.}

\item{\code{layers}}{This field is a list of the layers of the network, you can use
subsetting to inspect them.}
}}


