% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\docType{class}
\name{PerceptronLayer-class}
\alias{PerceptronLayer-class}
\alias{perceptronLayer}
\title{The PerceptronLayer class, that implements a layer of Perceptron neurons.}
\arguments{
\item{input}{The actual data to be fed to the layer, this input's
dimentions vary with the chosen \code{n}.}

\item{ins}{The list of vectors of inputs to the first layer in the network}

\item{outs}{The list of vectors of outputs of the last layer in the network}

\item{epochs}{How many rounds of training to run}

\item{tax}{This is the learning rate, aka eta}

\item{maxErr}{A contition to early stop the training process}
}
\value{
The computed value using the Perceptron model.

Vector of computed values of the same size of the last layer
}
\description{
The PerceptronLayer class, that implements a layer of Perceptron neurons.
}
\section{Fields}{

\describe{
\item{\code{n}}{The number of neurons to create in the layer}

\item{\code{dims}}{A vector of dimensions of the inputs to the layer}

\item{\code{neurons}}{A list with the internal neurons}
}}


